requirements.txt
Pillow==6.0.0
PyAutoGUI==0.9.45
PyGetWindow==0.0.6
PyMsgBox==1.0.7
PyRect==0.1.4
PyScreeze==0.1.22
python-decouple==3.1
PyTweening==1.0.3
selenium==3.141.0
urllib3==1.25.3
